ENH: Working through warnings in the Code/Common directory.

+++ Modules/Core/Common/include/itkOctreeNode.h
@@ -72,7 +72,7 @@ public:
    * member function.  Behavior is undefined when the child is another Octree.
    * \return A value between 0 and 255 to indicate the color of the Desired child.
    */
-  int GetColor(void) const;
+  long int GetColor(void) const;
 
   /**
    * Sets the color value of the specified Child for this OctreeNode
